Refine CloneAction._update()
The process of updating the branch to the most recent version contained
in the upstream repository wasn't quite right: changes ended up being
left over rather than discarded as they should.  This change fixes the
issue by forcing the "git reset --hard" to reset to the "origin" version
of the branch.