Use the actual user login
If no organization or user is specified as part of the repository name,
we were substituting the user's login from the command line.  However,
this isn't always a match; in particular, if the login name contains a
".", Github converts that to a "-".  To correct for this logic, this
change asks Github for what it thinks the user's login is, then uses
that to construct the repository's full name.